pyproject.toml
[tool.poetry]
name = "my_agent"
version = "0.1.0"
description = "Advanced LangGraph agent with multiple tools"
authors = ["langchain-ai"]
packages = [{ include = "my_agent" }]

[tool.poetry.dependencies]
python = ">=3.9.0,<3.13"
langgraph = "^0.2.0"
langchain-core = ">=0.3.17,<0.4.0"
langchain-openai = ">=0.2.0"
langchain-community = ">=0.3.0"
langchain-experimental = ">=0.0.49"
openai = "^1.12.0"
matplotlib = "^3.8.0"
python-dotenv = "^1.0.0"
tavily-python = "^0.3.0"
pydantic = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"